Move fetch-related tasks into fetching_tasks
