+++ dags/harvest_dag.py
@@ -7,8 +7,7 @@ from airflow.models.param import Param
 from typing import Optional
 
 
-from rikolti.dags.shared_tasks.shared_tasks import create_vernacular_version_task
-from rikolti.dags.shared_tasks.shared_tasks import fetch_collection_task
+from rikolti.dags.shared_tasks.fetching_tasks import fetching_tasks
 from rikolti.dags.shared_tasks.shared_tasks import get_registry_data_task
 from rikolti.dags.shared_tasks.shared_tasks  import create_mapped_version_task
 from rikolti.dags.shared_tasks.shared_tasks  import map_page_task
@@ -88,15 +87,6 @@ def get_mapped_page_filenames_task(mapped_page_batches: list[list[dict]]):
 
     return batches
 
-
-@task_group(group_id='fetching')
-def fetching_tasks(collection: Optional[dict] = None):
-    vernacular_version = create_vernacular_version_task(
-        collection=collection['registry_fetchdata'])
-    fetched_page_batches = fetch_collection_task(
-        collection=collection['registry_fetchdata'], 
-        vernacular_version=vernacular_version)
-    return fetched_page_batches
 
 
 @task_group(group_id='mapping')
